Amended parsing of data for Version 23

+++ TempoDisc/TempoDisc/Utilities/Helper/TDHelper.m
@@ -63,7 +63,7 @@
 																					[humArrayBytes addObject:[NSString stringWithFormat:@"%hhu", byte]];
 																				}
 																				//REDO
-																				NSLog(@"Error : %@ -> Index : %hd", error, indexDownload);
+																				NSLog(@"Error : %@ -> Index : %ld", error, (long)indexDownload);
 																				[self reWriteAndRead:peripheral];
 																			}];
 												  }];
